Add files via upload
The 3Dv2 and the Simulator3D are the main scripts, each one of this pieces of code is a separated program. I finished them after taking the Self-driving car Udacity free course. The 3Dv2 process in first place the algorithms and show graphics instead. The other one show you "live" the process of driving.